[FastRoute] Use long for wire length

+++ src/FastRoute/src/GlobalRouter.cpp
@@ -1709,7 +1709,7 @@ GlobalRouter::ROUTE_ GlobalRouter::getRoute()
 
 void GlobalRouter::computeWirelength()
 {
-  int totalWirelength = 0;
+  long totalWirelength = 0;
   for (auto &net_route : _routes) {
     GRoute &route = net_route.second;
     for (GSegment &segment : route) {
@@ -1723,7 +1723,7 @@ void GlobalRouter::computeWirelength()
     }
   }
   std::cout << std::fixed << "[INFO] Total wirelength: "
-            << (int) totalWirelength / _grid->getDatabaseUnit() << " um\n";
+            << (long) totalWirelength / _grid->getDatabaseUnit() << " um\n";
 }
 
 void GlobalRouter::mergeSegments()
